requirements.txt
groq
crewai==0.28.6
streamlit==1.30.0
python-dotenv==1.0.0
requests==2.31.0